requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
flower==1.0.0
humanize==3.13.1
kombu==5.2.3
mysqlclient==2.1.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
PyJWT==1.7.1
pymongo==3.12.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.2.4
tornado==6.1
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
